requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
google-genai>=0.1.0
pydantic>=2.10.0
python-dotenv>=1.0.1
